pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbus"
version = "0.1.0"
description = "GitHub-native agent state coordination layer"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["agents", "github", "coordination", "ai", "multi-agent"]
dependencies = [
    "requests>=2.28",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock", "responses"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentbus*"]
